requirements.txt
Flask
selenium
pymongo
requests
python-dotenv
